#515 move mast form to new component
